+++ CHANGELOG.md
@@ -40,6 +40,7 @@ the detailed section referring to by linking pull requests or issues.
 * Implement ContractDefinition service for Data Management API (#940)
 * Implement ContractNegotiation service for Data Management API (#957)
 * In-memory implementation of PolicyStore (#930)
+* Implement AssetLoader, AssetIndex, DataAddressResolver for SQL (#863)
 
 #### Changed